Initial Commit
Instantiated a new virtual environment (AlgoPy).
Added a script which runs Djikstra's algorithm on a graph to
determine the shortest path from a given starting node, to any
other node. Does not return the shortest path from a given
start to a given terminal but instead returns a dictionary
listing vertices, predecessor and path weight.

Signed-off-by: Aaron Zolotor <azolotor@tuta.io>

 Changes to be committed:
	new file:   .gitignore
	new virtual_environment:   AlgoPy/bin/activate
	new file:   djikstra.py